:zap: Moves item reusable methods into mixin
